// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/food-nutrition-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FoodNutritionItemHttpResponse } from './food-nutrition-item-http-response';

export interface FoodNutritionHttpResponse {
    alcohol?: number;
    calories?: number;
    carbs?: number;
    fat?: number;
    fiber?: number;
    items?: Array<FoodNutritionItemHttpResponse>;
    notes?: string | null;
    protein?: number;
}