// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/food-recognition-job-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FoodNutritionHttpResponse } from './food-nutrition-http-response';
import { FoodRecognitionImageHttpResponse } from './food-recognition-image-http-response';
import { FoodVisionHttpResponse } from './food-vision-http-response';

export interface FoodRecognitionJobHttpResponse {
    additionalImages?: Array<FoodRecognitionImageHttpResponse> | null;
    createdOnUtc?: string;
    description?: string | null;
    errorCode?: string | null;
    id?: string;
    imageAssetId?: string;
    imageUrl?: string;
    isProductLabel?: boolean;
    nutrition?: FoodNutritionHttpResponse | null;
    nutritionErrorCode?: string | null;
    status?: string;
    updatedOnUtc?: string;
    vision?: FoodVisionHttpResponse | null;
}