// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/food-vision-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FoodVisionItemHttpModel } from './food-vision-item-http-model';
import { ProductLabelHttpResponse } from './product-label-http-response';

export interface FoodVisionHttpResponse {
    items?: Array<FoodVisionItemHttpModel>;
    notes?: string | null;
    productLabel?: ProductLabelHttpResponse | null;
}