// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/product-label-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ProductLabelHttpResponse {
    alcohol?: number | null;
    baseAmount?: number | null;
    baseUnit?: string | null;
    brand?: string | null;
    calories?: number | null;
    carbs?: number | null;
    fat?: number | null;
    fiber?: number | null;
    name?: string | null;
    notes?: string | null;
    protein?: number | null;
}