// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/recipe-import-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecipeImportIngredientHttpResponse } from './recipe-import-ingredient-http-response';

export interface RecipeImportHttpResponse {
    authorNutrition?: string | null;
    cookMinutes?: number | null;
    description?: string | null;
    ingredients?: Array<RecipeImportIngredientHttpResponse>;
    name?: string;
    prepMinutes?: number | null;
    servings?: number | null;
    sourceUrl?: string | null;
    steps?: Array<string>;
}